// src/services/sited_io/media/v1/media_pb.ts
export class MediaResponse extends Message<MediaResponse> {
  /**
   * @generated from field: string media_id = 1;
   */
  mediaId = "";

  /**
   * @generated from field: repeated string offer_ids = 2;
   */
  offerIds: string[] = [];

  /**
   * @generated from field: string shop_id = 3;
   */
  shopId = "";

  /**
   * @generated from field: string user_id = 4;
   */
  userId = "";

  /**
   * @generated from field: int64 created_at = 5;
   */
  createdAt = protoInt64.zero;

  /**
   * @generated from field: int64 updated_at = 6;
   */
  updatedAt = protoInt64.zero;

  /**
   * @generated from field: string name = 7;
   */
  name = "";

  /**
   * @generated from field: string file_name = 8;
   */
  fileName = "";

  /**
   * @generated from field: int64 ordering = 9;
   */
  ordering = protoInt64.zero;

  constructor(data?: PartialMessage<MediaResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.media.v1.MediaResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "media_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "offer_ids", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
    { no: 3, name: "shop_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "user_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "created_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 6, name: "updated_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 7, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 8, name: "file_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 9, name: "ordering", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MediaResponse {
    return new MediaResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MediaResponse {
    return new MediaResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MediaResponse {
    return new MediaResponse().fromJsonString(jsonString, options);
  }

  static equals(a: MediaResponse | PlainMessage<MediaResponse> | undefined, b: MediaResponse | PlainMessage<MediaResponse> | undefined): boolean {
    return proto3.util.equals(MediaResponse, a, b);
  }
}